OR-1291 update code structure
